split out checkout item into separate component, done with styling as well
