+++ src/components/checkout-item/checkout-item.scss
@@ -1,0 +1,51 @@
+.checkout_item_container {
+  width: 100%;
+  min-height: 100px;
+  padding: 15px 0;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  border-bottom: 1px solid darkgrey;
+
+  .image-container {
+    width: 23%;
+    padding-right: 15px;
+
+    img {
+      width: 100%;
+      height: 100%;
+    }
+  }
+
+  .name,
+  .quantity,
+  .price {
+    width: 23%;
+  }
+
+  .quantity,
+  .price {
+    font-size: 18px;
+  }
+
+  .quantity {
+    display: flex;
+
+    .value {
+      margin: 0px 10px 5px;
+    }
+  }
+
+  .remove-button {
+    width: 8%;
+    padding-left: 12px;
+  }
+}
+
+button {
+  border: none;
+  background: none;
+  font-size: 18px;
+  cursor: pointer;
+}
