+++ src/routes/checkout-page/checkout-page.scss
@@ -15,30 +15,13 @@
     border-bottom: 1px solid darkgrey;
 
     .header_block {
+      font-size: 18px;
       text-transform: capitalize;
       width: 23%;
 
       &:last-child {
         width: 8%;
       }
-    }
-  }
-
-  .checkout_item {
-    width: 900px;
-    margin: 15px auto;
-    padding: 15px 30px;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-
-    img {
-      width: 20%;
-    }
-
-    button {
-      margin: 0 15px;
     }
   }
